+++ mods/BrewLAN/modules/BrewUI/spawnmenu/hook/lua/SimCallbacks.lua
@@ -21,7 +21,7 @@ local function SpawnUnitMesh(id, x, y, z, pitch, yaw, roll)
     end
 end
 
-local function SetWorldCameraToUnitIconAngle(location)
+local function SetWorldCameraToUnitIconAngle(location, zoom)
     local sx = 1/6
     table.insert( Sync.CameraRequests, {
         Name = 'WorldCamera',
@@ -29,7 +29,7 @@ local function SetWorldCameraToUnitIconAngle(location)
         Marker = {
             orientation = VECTOR3(math.pi*(1+sx), math.pi*sx, 0),
             position = location,
-            zoom = FLOAT(30),
+            zoom = FLOAT(zoom),
         },
         HeadingRate = 0,
         Callback = {
@@ -54,18 +54,15 @@ Callbacks.BoxFormationSpawn = function(data)
         axe = axe == 'x' and 'SizeX' or 'SizeZ' --local axes = {x='SizeX', z='SizeZ'}
         return unitbp.Footprint
         and unitbp.Footprint[axe]
-        or unitbp[axe]
-        or 1
+        or math.ceil(unitbp[axe] or 1)
     end
 
     local function RoundToSkirt(axe, val)
-        return unitbp.Physics.MotionType ~= 'RULEUMT_None'
-        and val
-        or math.floor(val) + (math.mod(FootprintSize(axe),2) == 1 and 0.5 or 0)
+        local alignment = math.mod(FootprintSize(axe),2) == 1
+        return unitbp.Physics.MotionType ~= 'RULEUMT_None' and val
+        or math.floor(val + (alignment and 0 or 0.5)) + (alignment and 0.5 or 0)
     end
 
-    local posX = math.floor(data.pos[1])
-    local posZ = math.floor(data.pos[3])
     local offsetX = unitbp.SizeX or 1
     local offsetZ = unitbp.SizeZ or 1
 
@@ -82,17 +79,25 @@ Callbacks.BoxFormationSpawn = function(data)
     local yaw = (data.yaw or 0) / 57.295779513
 
     for i = 1, data.count do
-        local x = RoundToSkirt('x', posX - startOffsetX + math.mod(i,squareX) * offsetX)
-        local z = RoundToSkirt('z', posZ - startOffsetZ + math.mod(math.floor(i/squareX), squareZ) * offsetZ)
+        local x = RoundToSkirt('x', (data.pos[1]) - startOffsetX + math.mod(i,squareX) * offsetX)
+        local z = RoundToSkirt('z', (data.pos[3]) - startOffsetZ + math.mod(math.floor(i/squareX), squareZ) * offsetZ)
         local unit = not data.MeshOnly and CreateUnitHPR(data.bpId, data.army, x, GetTerrainHeight(x,z), z, 0, yaw, 0)--blueprint, army, x, y, z, pitch, yaw, roll
         or SpawnUnitMesh(data.bpId, x, GetTerrainHeight(x,z), z, 0, yaw, 0)
 
         if unit.SetVeterancy then unit:SetVeterancy(data.veterancy) end
-        if data.CreateTarmac and unit.CreateTarmac and __blueprints[data.bpId].Display and __blueprints[data.bpId].Display.Tarmacs then
+        if data.CreateTarmac and unit.CreateTarmac and unitbp.Display and unitbp.Display.Tarmacs then
             unit:CreateTarmac(true,true,true,false,false)
         end
         if data.count == 1 and data.UnitIconCameraMode then
-            SetWorldCameraToUnitIconAngle{x, GetTerrainHeight(x,z), z}
+            local size = math.max(
+                (unitbp.SizeX or 1),
+                (unitbp.SizeY or 1) * 3,
+                (unitbp.SizeZ or 1),
+                (unitbp.Physics.SkirtSizeX or 1),
+                (unitbp.Physics.SkirtSizeZ or 1)
+            ) + math.abs(unitbp.CollisionOffsetY or 0)
+            local dist = size / math.tan(60 --[[* (9/16)]] * 0.5 * ((math.pi*2)/360))
+            SetWorldCameraToUnitIconAngle({x, GetTerrainHeight(x,z), z}, dist)
         end
     end
 end